requirements.txt
pandas>=2.0.0
langgraph>=0.2.0
google-genai>=1.0.0
python-dotenv>=1.0.0
